feat: implement Dashboard component with widgets and charts for admin and employee views
